added the entities in repos

+++ src/main/java/com/example/security/user/UserEntity.java
@@ -23,7 +23,8 @@ public class UserEntity implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
+    private Long id;
+
     private String firstName;
     private String lastName;
     private String email;
@@ -66,4 +67,5 @@ public class UserEntity implements UserDetails {
     public boolean isEnabled() {
         return true;
     }
+
 }
